requirements.txt
anyio==4.5.2
APScheduler==3.11.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
command-not-found==0.3
importlib-resources==6.4.5
iniconfig==2.0.0
joblib==1.4.2
keyring==18.0.1
kiwisolver==1.4.7
launchpadlib==1.10.13
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.7.5
nemo-emblems==5.2.0
netaddr==0.7.19
netifaces==0.10.4
nltk==3.9.1
numpy==1.24.4
oauthlib==3.1.0
onboard==1.4.1
packaging==20.3
PAM==0.4.2
pandas==2.0.3
pbr==6.1.0
pexpect==4.6.0
Pillow==7.0.0
pluggy==1.5.0
protobuf==3.6.1
psutil==5.5.1
pycairo==1.16.2
pycodestyle==2.12.1
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0.2
PyGObject==3.36.0
PyICU==2.4.2
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.6
pyparted==3.11.2
pyRFC3339==1.1
pytest==8.3.4
python-dateutil==2.9.0.post0
python-magic==0.4.16
python-telegram-bot==21.6
python-xapp==2.2.1
python-xlib==0.23
pytz==2024.2
pyxdg==0.26
PyYAML==6.0.2
random-word==1.0.13
requests==2.32.3
requests-file==1.4.3
requests-unixsocket==0.2.0
schedule==1.2.2
ssh-import-id==5.10
systemd-python==234
testresources==2.0.1
textblob==0.18.0.post0
typing-extensions==4.12.2
